00_mvtec_sample/02_remote_debug/remote_debug/requirements.txt
mvtec-halcon==22111.0.0
